perf: retain native mxfp4 weights on gpu

+++ converter/k3x_converter/mxfp4_cuda.py
@@ -56,4 +56,4 @@ def mxfp4_matvec(
         rows,
         columns,
     )
-    return output.to(torch.bfloat16)
+    return output
